requirements.txt
django
djangorestframework
django-rest-auth
requests